Unit tested slack::default cleaned for foodcritic
The LWRP still needs to be tested, but since I'll probably make a fake
cookbook to consume the LWRP in order to test it, I wanted to get the
base of the unit testing up first.

The versions are specified in a constant in the spec_helper, which will
be used to ensure all versions we care about are under test.

I also added rake tasks for Foodcritic and ChefSpec.  The default task
runs both.  I was able to remove all foodcritic issues except for this
line:
https://github.com/risk-io/chef-slack/blob/master/recipes/default.rb#L24.
I filed and issue with foodcritic to resolve it (acrmp/foodcritic#339).